--- debezium-core/src/main/java/io/debezium/transforms/debezium-timestamp-converter/src/main/java/org/telmate/SMT/OptionalTypeFieldSchema.java
@@ -1,13 +1,0 @@
-package org.telmate.SMT;
-
-import org.apache.kafka.connect.data.Schema;
-
-/**
- * Created by svegesna on 9/29/17.
- */
-public interface OptionalTypeFieldSchema {
-    /**
-     * @return the optional schema.
-     */
-    Schema optionalSchemaNonTimestamp();
-}
